Share changes sync logic across CLI and GUI
Move remote configuration and GitHub sync entry points into shared crates
without changing CLI behavior.

Change: factor-changes-sync-into-josh-github-changes
