+++ forges/josh-github-changes/Cargo.toml
@@ -11,12 +11,19 @@ keywords = ["git", "github", "monorepo", "workflow"]
 [dependencies]
 gix-hash.workspace = true
 anyhow.workspace = true
-url.workspace = true
+reqwest.workspace = true
+reqwest-middleware.workspace = true
 serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+url.workspace = true
+
 josh-changes.workspace = true
 josh-core.workspace = true
 josh-git-serde.workspace = true
+josh-github-auth.workspace = true
 josh-github-graphql.workspace = true
+josh-github-keyring.workspace = true
 josh-github-webhooks.workspace = true
 
 [dev-dependencies]
