+++ forges/josh-github-changes/src/lib.rs
@@ -1,11 +1,13 @@
 pub mod admission;
 mod cache;
 mod comments;
+pub mod connection;
 mod display;
 pub mod layout;
 mod node_ids;
 mod prs;
 pub mod repo;
+pub mod sync;
 
 pub use cache::{read_sync_fingerprint, store_sync_fingerprint, SyncFingerprint};
 pub use comments::{
