+++ josh-cli/src/forge/github/mod.rs
@@ -1,5 +1,3 @@
-pub mod cache;
-pub mod changes;
 mod login;
 
 pub use login::{api_connection_hint, login, logout, make_api_connection};
